cdn2nsp: verify tik sig + write fixed tik.
The modified ticket is only written if it was originally tampered in the first place. This is done to make sure it matches a tampered nxdumptool ticket.

Furthermore, an additional check is now in place to make sure NSPs are not generated for updates and DLC updates with tampered tickets, since there's no reason to modify them.